requirements.txt
streamlit==1.37.1
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.8
python-dotenv==1.0.1
unstructured==0.10.30
requests==2.32.3
beautifulsoup4==4.12.3
html5lib==1.1
tiktoken==0.7.0
faiss-cpu==1.8.0.post1
openai==1.109.1
nltk==3.9.1